Refactor PermitLOA type, refactor LOA endpoints to different file, add helper methods for comparing arrays of objects
