+++ frontend/src/common/helpers/equality.ts
@@ -4,19 +4,19 @@ import { Nullable } from "../types/common";
  * Check if two nullable values are different.
  * @param val1 First nullable value to be compared
  * @param val2 Second nullable value to be compared
- * @returns boolean value indicating if values are different.
+ * @returns true when only one of the values are empty, or both are non-empty and different, false otherwise
  */
 export const areValuesDifferent = <T>(
   val1?: Nullable<T>,
   val2?: Nullable<T>,
 ): boolean => {
-  if (!val1 && !val2) return false; // both empty === equal
+  if (!val1 && !val2) return false; // Both empty implicitly means that values are the same
 
   if ((val1 && !val2) || (!val1 && val2) || (val1 && val2 && val1 !== val2)) {
-    return true; // one empty, or both non-empty but different
+    return true; // Only one empty, or both are non-empty but different means that values are different
   }
 
-  return false; // values are equal otherwise
+  return false; // Values are considered equal otherwise
 };
 
 /**
@@ -40,5 +40,38 @@ export const doUniqueArraysHaveSameItems = <T extends (number | string)>(
     if (!set1.has(val)) return false;
   }
   
+  return true;
+};
+
+/**
+ * Determine whether or not two arrays, each with objects of a certain type identifiable by keys,
+ * have the same objects.
+ * @param arr1 First array consisting of identifiable objects
+ * @param arr2 Second array consisting of identifiable objects
+ * @param key Function that returns the identifier of an object of the given type
+ * @param equalFn Function that compares equality of two objects of the given type
+ * @returns Whether or not the two arrays have the same objects
+ */
+export const doUniqueArraysHaveSameObjects = <T, K extends (number | string)>(
+  arr1: T[],
+  arr2: T[],
+  key: (item: T) => K,
+  equalFn: (item1: T, item2: T) => boolean,
+) => {
+  const map1 = new Map<K, T>(arr1.map(item => [key(item), item]));
+  const map2 = new Map<K, T>(arr2.map(item => [key(item), item]));
+
+  for (const [key, item] of map1) {
+    const itemInOtherMapWithSameKey = map2.get(key);
+    if (!itemInOtherMapWithSameKey || !equalFn(item, itemInOtherMapWithSameKey))
+      return false;
+  }
+
+  for (const [key, item] of map2) {
+    const itemInOtherMapWithSameKey = map1.get(key);
+    if (!itemInOtherMapWithSameKey || !equalFn(item, itemInOtherMapWithSameKey))
+      return false;
+  }
+
   return true;
 };
