+++ frontend/src/features/permits/hooks/useInitApplicationFormData.ts
@@ -7,7 +7,7 @@ import { BCeIDUserDetailContext } from "../../../common/authentication/OnRouteBC
 import { CompanyProfile } from "../../manageProfile/types/manageProfile";
 import { Nullable } from "../../../common/types/common";
 import { PermitType } from "../types/PermitType";
-import { LOADetail } from "../../settings/types/SpecialAuthorization";
+import { LOADetail } from "../../settings/types/LOADetail";
 import { applyUpToDateLOAsToApplication } from "../helpers/permitLOA";
 import { getDefaultValues } from "../helpers/getDefaultApplicationFormData";
 import { applyLCVToApplicationData } from "../helpers/permitLCV";
@@ -15,6 +15,7 @@ import { PowerUnit, Trailer } from "../../manageVehicles/types/Vehicle";
 import { getIneligibleSubtypes } from "../helpers/permitVehicles";
 import { PermitCondition } from "../types/PermitCondition";
 import { EMPTY_VEHICLE_DETAILS, PermitVehicleDetails } from "../types/PermitVehicleDetails";
+import { PermitLOA } from "../types/PermitLOA";
 
 /**
  * Custom hook for populating the form using fetched application data, as well as current company id and user details.
@@ -30,7 +31,7 @@ import { EMPTY_VEHICLE_DETAILS, PermitVehicleDetails } from "../types/PermitVehi
 export const useInitApplicationFormData = (
   permitType: PermitType,
   isLcvDesignated: boolean,
-  loas: LOADetail[],
+  companyLOAs: LOADetail[],
   inventoryVehicles: (PowerUnit | Trailer)[],
   companyInfo: Nullable<CompanyProfile>,
   applicationData?: Nullable<Application>,
@@ -56,7 +57,7 @@ export const useInitApplicationFormData = (
         ),
         isLcvDesignated,
       ),
-      loas,
+      companyLOAs,
       inventoryVehicles,
       ineligiblePowerUnitSubtypes,
       ineligibleTrailerSubtypes,
@@ -67,7 +68,7 @@ export const useInitApplicationFormData = (
     applicationData,
     userDetails,
     isLcvDesignated,
-    loas,
+    companyLOAs,
     inventoryVehicles,
   ]);
 
@@ -114,7 +115,7 @@ export const useInitApplicationFormData = (
     });
   };
 
-  const onUpdateLOAs = (updatedLOAs: LOADetail[]) => {
+  const onUpdateLOAs = (updatedLOAs: PermitLOA[]) => {
     setValue("permitData.loas", updatedLOAs);
   };
