+++ frontend/src/features/permits/pages/Application/components/form/PermitForm.tsx
@@ -6,8 +6,8 @@ import { ApplicationDetails } from "../../../../components/form/ApplicationDetai
 import { ContactDetails } from "../../../../components/form/ContactDetails";
 import { PermitDetails } from "./PermitDetails";
 import { VehicleDetails } from "./VehicleDetails/VehicleDetails";
-import { PermitLOA } from "./PermitLOA";
-import { LOADetail } from "../../../../../settings/types/SpecialAuthorization";
+import { PermitLOASection } from "./PermitLOASection";
+import { PermitLOA } from "../../../../types/PermitLOA";
 import { isVehicleSubtypeLCV } from "../../../../../manageVehicles/helpers/vehicleSubtypes";
 import { getStartOfDate } from "../../../../../../common/helpers/formatDate";
 import { useApplicationFormContext } from "../../../../hooks/useApplicationFormContext";
@@ -50,7 +50,7 @@ export const PermitForm = () => {
   const expiryDate = formData.permitData.expiryDate;
   const permitConditions = formData.permitData.commodities;
   const vehicleFormData = formData.permitData.vehicleDetails;
-  const currentSelectedLOAs = formData.permitData.loas as LOADetail[];
+  const currentSelectedLOAs = formData.permitData.loas as PermitLOA[];
 
   return (
     <Box className="permit-form layout-box">
@@ -70,7 +70,7 @@ export const PermitForm = () => {
 
         <ContactDetails feature={feature} />
 
-        <PermitLOA
+        <PermitLOASection
           permitType={permitType}
           startDate={startDate}
           selectedLOAs={currentSelectedLOAs}
