+++ frontend/src/features/permits/types/PermitData.ts
@@ -4,8 +4,8 @@ import { Nullable } from "../../../common/types/common";
 import { PermitContactDetails } from "./PermitContactDetails";
 import { PermitVehicleDetails } from "./PermitVehicleDetails";
 import { PermitMailingAddress } from "./PermitMailingAddress";
-import { LOADetail } from "../../settings/types/SpecialAuthorization";
 import { PermitCondition } from "./PermitCondition";
+import { PermitLOA } from "./PermitLOA";
 
 export interface PermitData {
   startDate: Dayjs;
@@ -19,5 +19,5 @@ export interface PermitData {
   companyName?: Nullable<string>;
   doingBusinessAs?: Nullable<string>;
   clientNumber?: Nullable<string>;
-  loas?: Nullable<LOADetail[]>;
+  loas?: Nullable<PermitLOA[]>;
 }
