+++ frontend/src/features/settings/types/SpecialAuthorization.ts
@@ -1,6 +1,4 @@
-import { doUniqueArraysHaveSameItems } from "../../../common/helpers/equality";
-import { Nullable, RequiredOrNull } from "../../../common/types/common";
-import { PermitType } from "../../permits/types/PermitType";
+import { RequiredOrNull } from "../../../common/types/common";
 
 export const NO_FEE_PERMIT_TYPES = {
   CA_GOVT: "CA_GOVT",
@@ -29,69 +27,6 @@ export const noFeePermitTypeDescription = (noFeePermitType: NoFeePermitType) => 
   }
 };
 
-export interface LOADetail {
-  loaId: number;
-  loaNumber: number;
-  companyId: number;
-  startDate: string;
-  expiryDate?: Nullable<string>;
-  documentId: string;
-  fileName: string;
-  loaPermitType: PermitType[];
-  comment?: Nullable<string>;
-  powerUnits: string[];
-  trailers: string[];
-  originalLoaId: number;
-  previousLoaId?: Nullable<number>;
-}
-
-export interface CreateLOARequestData {
-  startDate: string;
-  expiryDate?: Nullable<string>;
-  loaPermitType: PermitType[];
-  // document: Buffer;
-  comment?: Nullable<string>;
-  powerUnits: string[];
-  trailers: string[];
-}
-
-export interface UpdateLOARequestData {
-  startDate: string;
-  expiryDate?: Nullable<string>;
-  loaPermitType: PermitType[];
-  // document?: Buffer;
-  comment?: Nullable<string>;
-  powerUnits: string[];
-  trailers: string[];
-}
-
-/**
- * Determine whether or not two LOAs have the same details.
- * @param loa1 First LOA
- * @param loa2 Second LOA
- * @returns Whether or not the two LOAs have the same details
- */
-export const areLOADetailsEqual = (
-  loa1?: Nullable<LOADetail>,
-  loa2?: Nullable<LOADetail>,
-) => {
-  if (!loa1 && !loa2) return true;
-  if (!loa1 || !loa2) return false;
-  
-  return loa1.loaId === loa2.loaId
-    && loa1.loaNumber === loa2.loaNumber
-    && loa1.companyId === loa2.companyId
-    && loa1.startDate === loa2.startDate
-    && loa1.expiryDate === loa2.expiryDate
-    && loa1.documentId === loa2.documentId
-    && loa1.fileName === loa2.fileName
-    && doUniqueArraysHaveSameItems<string>(loa1.loaPermitType, loa2.loaPermitType)
-    && loa1.comment === loa2.comment
-    && doUniqueArraysHaveSameItems<string>(loa1.powerUnits, loa2.powerUnits)
-    && doUniqueArraysHaveSameItems<string>(loa1.trailers, loa2.trailers)
-    && loa1.originalLoaId === loa2.originalLoaId
-    && loa1.previousLoaId === loa2.previousLoaId;
-};
 export interface SpecialAuthorizationData {
   companyId: number;
   specialAuthId: number;
